Refacto controllers
