+++ app/Controllers/AuthenticationsController.php
@@ -3,17 +3,18 @@
 namespace App\Controllers;
 
 use App\Models\User;
+use Core\Http\Controllers\Controller;
 use Core\Http\Request;
 use Lib\Authentication\Auth;
 use Lib\FlashMessage;
 
-class AuthenticationsController
+class AuthenticationsController extends Controller
 {
-    private string $layout = 'login';
+    protected string $layout = 'login';
 
     public function new(): void
     {
-        $this->render('new');
+        $this->render('authentications/new');
     }
 
     public function authenticate(Request $request): void
@@ -37,23 +38,5 @@ class AuthenticationsController
         Auth::logout();
         FlashMessage::success('Logout realizado com sucesso!');
         $this->redirectTo(route('users.login'));
-    }
-
-
-    /**
-     * @param array<string, mixed> $data
-     */
-    private function render(string $view, array $data = []): void
-    {
-        extract($data);
-
-        $view = '/var/www/app/views/authentications/' . $view . '.phtml';
-        require '/var/www/app/views/layouts/' . $this->layout . '.phtml';
-    }
-
-    private function redirectTo(string $location): void
-    {
-        header('Location: ' . $location);
-        exit;
     }
 }
